+++ CHANGELOG.md
@@ -4,3 +4,4 @@
   type but that allows only simple syntax shapes to be described.
 - Rename `Flag` field `Arg` to `Shape` and change it's type from `string` to `syntaxshape.SyntaxShape`.
 - Change `PositionalArg` field `Shape` type from `string` to `syntaxshape.SyntaxShape`.
+- Rename `PluginSignature` field `Usage` to `Desc` and `UsageEx` to `Description`.